removing segment and adding it to the woopra controller
