+++ pkg/controller/woopra.go
@@ -10,6 +10,10 @@ type woopraResponse struct {
 	WoopraDomain string `json:"woopra-domain"`
 }
 
+type segmentResponse struct {
+	SegmentWriteKey string `json:"segment-write-key"`
+}
+
 // Woopra implements the segment endpoint, which is invoked to
 // retrieve the woopra domain for the ui.
 type Woopra struct {
@@ -27,4 +31,10 @@ func NewWoopra(config configuration.Configuration) *Woopra {
 func (w *Woopra) GetWoopraDomain(ctx *gin.Context) {
 	domain := w.config.GetWoopraDomain()
 	ctx.JSON(http.StatusOK, woopraResponse{WoopraDomain: domain})
+}
+
+// GetSegmentWriteKey returns segment-write-key content for UI.
+func (s *Woopra) GetSegmentWriteKey(ctx *gin.Context) {
+	segmentWriteKey := s.config.GetSegmentWriteKey()
+	ctx.JSON(http.StatusOK, segmentResponse{SegmentWriteKey: segmentWriteKey})
 }
